Move Python bindings to a submodule in the main crate
