+++ umbral-pre/src/lib.rs
@@ -107,10 +107,16 @@
 // Allows us to mark items in the documentation as gated under specific features.
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+#[cfg(feature = "std")]
+extern crate std;
+
 extern crate alloc;
 
 #[cfg(feature = "bench-internals")]
 pub mod bench; // Re-export some internals for benchmarks.
+
+#[cfg(feature = "bindings-python")]
+pub mod bindings_python;
 
 mod capsule;
 mod capsule_frag;
